[AXON-1487] Move rovodev-only util code to rovo-dev/util
