+++ src/rovo-dev/rovoDevProcessManager.ts
@@ -5,9 +5,9 @@ import net from 'net';
 import packageJson from 'package.json';
 import path from 'path';
 import { RovoDevLogger } from 'src/logger';
-import { downloadAndUnzip } from 'src/util/downloadFile';
-import { getFsPromise } from 'src/util/fsPromises';
-import { waitFor } from 'src/util/waitFor';
+import { downloadAndUnzip } from 'src/rovo-dev/util/downloadFile';
+import { getFsPromise } from 'src/rovo-dev/util/fsPromises';
+import { waitFor } from 'src/rovo-dev/util/waitFor';
 import { Disposable, Event, EventEmitter, ExtensionContext, Terminal, Uri, window, workspace } from 'vscode';
 
 import { DetailedSiteInfo, isBasicAuthInfo, ProductJira } from '../atlclients/authInfo';
